backend/requirements.txt
# Core env/config
python-dotenv>=1.0.1,<2.0.0

# Vector database
qdrant-client>=1.10.1,<2.0.0

# Embeddings + retrieval
sentence-transformers>=3.0.1,<4.0.0
rank-bm25>=0.2.2,<1.0.0
numpy>=1.26.4,<3.0.0

# PDF extraction
PyMuPDF>=1.24.9,<2.0.0

# Graph database
neo4j>=5.23.1,<6.0.0

# LLM (new + legacy compatibility used in code)
openai>=1.30.0,<2.0.0
google-genai>=1.26.0,<2.0.0
google-generativeai>=0.7.2,<1.0.0

# Optional framework used by retriever (code falls back if missing)
agno>=1.0.0,<2.0.0

# Optional acceleration/runtime helpers commonly needed by sentence-transformers
torch>=2.3.1,<3.0.0
transformers>=4.44.2,<5.0.0
huggingface-hub>=0.24.6,<1.0.0


#fine tuning
datasets>=2.20.0,<3.0.0
peft>=0.11.1,<1.0.0
trl>=0.9.6,<1.0.0
accelerate>=0.33.0,<1.0.0
bitsandbytes>=0.43.1


fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.0
